+++ src/channel.ts
@@ -57,6 +57,7 @@ import {
   PollVoteData,
   SendMessageOptions,
   AscDesc,
+  PartialUpdateMemberAPIResponse,
 } from './types';
 import { Role } from './permissions';
 import { DEFAULT_QUERY_CHANNEL_MESSAGE_LIST_PAGE_SIZE } from './constants';
@@ -308,12 +309,12 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
    *
    * @return {Promise<ChannelMemberResponse<StreamChatGenerics>>} Updated member
    */
-  async partialUpdateMember(user_id: string, updates: PartialUpdateMember<StreamChatGenerics>) {
+  async partialUpdateMember(user_id: string, updates: PartialUpdateMember) {
     if (!user_id) {
       throw Error('Please specify the user id');
     }
 
-    return await this.getClient().patch<ChannelMemberResponse<StreamChatGenerics>>(
+    return await this.getClient().patch<PartialUpdateMemberAPIResponse<StreamChatGenerics>>(
       this._channelURL() + `/member/${encodeURIComponent(user_id)}`,
       updates,
     );
@@ -641,6 +642,94 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       channel_cid: this.cid,
       ...opts,
     });
+  }
+
+  /**
+   * archive - archives the current channel
+   * @param {{ user_id?: string }} opts user_id if called server side
+   * @return {Promise<ChannelMemberResponse<StreamChatGenerics>>} The server response
+   *
+   * example:
+   * await channel.archives();
+   *
+   * example server side:
+   * await channel.archive({user_id: userId});
+   *
+   */
+  async archive(opts: { user_id?: string } = {}) {
+    const cli = this.getClient();
+    const uid = opts.user_id || cli.userID;
+    if (!uid) {
+      throw Error('A user_id is required for archiving a channel');
+    }
+    const resp = await this.partialUpdateMember(uid, { set: { archived: true } });
+    return resp.channel_member;
+  }
+
+  /**
+   * unarchive - unarchives the current channel
+   * @param {{ user_id?: string }} opts user_id if called server side
+   * @return {Promise<ChannelMemberResponse<StreamChatGenerics>>} The server response
+   *
+   * example:
+   * await channel.unarchive();
+   *
+   * example server side:
+   * await channel.unarchive({user_id: userId});
+   *
+   */
+  async unarchive(opts: { user_id?: string } = {}) {
+    const cli = this.getClient();
+    const uid = opts.user_id || cli.userID;
+    if (!uid) {
+      throw Error('A user_id is required for unarchiving a channel');
+    }
+    const resp = await this.partialUpdateMember(uid, { set: { archived: false } });
+    return resp.channel_member;
+  }
+
+  /**
+   * pin - pins the current channel
+   * @param {{ user_id?: string }} opts user_id if called server side
+   * @return {Promise<ChannelMemberResponse<StreamChatGenerics>>} The server response
+   *
+   * example:
+   * await channel.pin();
+   *
+   * example server side:
+   * await channel.pin({user_id: userId});
+   *
+   */
+  async pin(opts: { user_id?: string } = {}) {
+    const cli = this.getClient();
+    const uid = opts.user_id || cli.userID;
+    if (!uid) {
+      throw Error('A user_id is required for pinning a channel');
+    }
+    const resp = await this.partialUpdateMember(uid, { set: { pinned: true } });
+    return resp.channel_member;
+  }
+
+  /**
+   * unpin - unpins the current channel
+   * @param {{ user_id?: string }} opts user_id if called server side
+   * @return {Promise<ChannelMemberResponse<StreamChatGenerics>>} The server response
+   *
+   * example:
+   * await channel.unpin();
+   *
+   * example server side:
+   * await channel.unpin({user_id: userId});
+   *
+   */
+  async unpin(opts: { user_id?: string } = {}) {
+    const cli = this.getClient();
+    const uid = opts.user_id || cli.userID;
+    if (!uid) {
+      throw Error('A user_id is required for unpinning a channel');
+    }
+    const resp = await this.partialUpdateMember(uid, { set: { pinned: false } });
+    return resp.channel_member;
   }
 
   /**
@@ -1079,6 +1168,8 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       }),
     };
 
+    this.getClient().polls.hydratePollCache(messageSet.messages, true);
+
     const areCapabilitiesChanged =
       [...(state.channel.own_capabilities || [])].sort().join() !==
       [...(Array.isArray(this.data?.own_capabilities) ? (this.data?.own_capabilities as string[]) : [])].sort().join();
@@ -1460,31 +1551,6 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
             hidden: event.channel?.hidden ?? channel.data?.hidden,
             own_capabilities: event.channel?.own_capabilities ?? channel.data?.own_capabilities,
           };
-        }
-        break;
-      case 'poll.updated':
-        if (event.poll) {
-          channelState.updatePoll(event.poll, event.message?.id || '');
-        }
-        break;
-      case 'poll.vote_casted':
-        if (event.poll_vote && event.poll) {
-          channelState.addPollVote(event.poll_vote, event.poll, event.message?.id || '');
-        }
-        break;
-      case 'poll.vote_changed':
-        if (event.poll_vote && event.poll) {
-          channelState.updatePollVote(event.poll_vote, event.poll, event.message?.id || '');
-        }
-        break;
-      case 'poll.vote_removed':
-        if (event.poll_vote && event.poll) {
-          channelState.removePollVote(event.poll_vote, event.poll, event.message?.id || '');
-        }
-        break;
-      case 'poll.closed':
-        if (event.message) {
-          channelState.addMessageSorted(event.message, false, false);
         }
         break;
       case 'reaction.new':